bugfix(media): Thumbnail loading
Fixes small bug where if you close a media player playing a certain song, and reopen it and continue, the thumbnail does not load because the title and artist are the same. Now, if media info is ever None, it wipes the last title and artist used for thumbnail detect.