feat(wifi): Add Ethernet connection icon
Adds the ability to show an ethernet connection icon when connected to ethernet, instead of showing 'no wifi' icon. Since `netsh` seems to give problems for the lan interfaces, I use the winsdk to identify whether the current connection is WiFi or not. If it is WiFi or disconnected, it uses the `netsh` code, otherwise sets an Ethernet icon.

I did switch to Segoe Fluent Icons because I could not get the NerdFont icons to look nice for the ethernet one.