fix(media): Correct hide/show on session change
Introduced an event listener for the session change to set the hide/show properly, instead of relying on the media_info change like before.

Known bug: I also identified that the 'C++' error fixed in commit `b08c77a` has resurfaced, due to us making our own event loop, for which the tasks are not killed. It would be better if the bar would call a 'on_close' method for all widgets, and we could put this kind of graceful stopping in there.