feat(launch): Do not update bar in foreground on launch
Previously, the bar would force update each widget before showing the UI. On slower systems (or heavier widgets), this could cause the bar to be invisible for some time, making it unclear for the user that yasb is running. Deferring the initial update to an instantaneous background call should help make the bar visible sooner.