fix(media): Fix error on config change
Fixes a bug where the bar crashes on config reload if the media widget is not enabled. This was caused by the bar initializing a new media manager in the shutdown loop because it was not made by any of the widgets. Fixed by only trying to stop it if an instance exists.